early int conversion bandaid

+++ src/main/cpp/infrastructure/PWMAngleSensor.cpp
@@ -2,7 +2,7 @@
 
 #include "infrastructure/ShuffleboardWidgets.h"
 
-AngleSensor::AngleSensor(int deviceID, double alignment) noexcept
+AngleSensor::AngleSensor(int deviceID, int alignment) noexcept
     : canCoder_(deviceID), alignment_(alignment) {
     // Configure the sensor range and other settings
     canCoder_.ConfigAbsoluteSensorRange(ctre::phoenix::sensors::AbsoluteSensorRange::Signed_PlusMinus180);
